requirements.txt
Flask 
Flask-CORS 
langchain-community 
langchain-anthropic 
SQLAlchemy
python-dotenv
mysql-connector-python
